feat: Update text for client basic authentication #2077
