+++ lib/dashboard/drawer/ssi/advanced_settings2/oidc4vc_settngs/view/oidc4vc_settings_menu.dart
@@ -43,7 +43,7 @@ class Oidc4vcSettingMenuView extends StatelessWidget {
           const SubjectSyntaxTypeWidget(),
           const CryptographicHolderBindingWidget(),
           const ScopeParameterWidget(),
-          const PreRegisteredWalletWidget(),
+          const ClientAuthenticationWidget(),
           DrawerItem(
             title: l10n.clientMetadata,
             onTap: () {
