+++ lib/dashboard/drawer/ssi/advanced_settings2/oidc4vc_settngs/widget/widget.dart
@@ -1,6 +1,6 @@
 export 'cryptograhic_holder_binding.dart';
 export 'did_key_type_widget.dart';
-export 'pre_registered_wallet_widget.dart';
+export 'client_authentication_widget.dart';
 export 'scope_parameter.dart';
 export 'security_level_widget.dart';
 export 'six_or_four_pin_widget.dart';
