+++ lib/l10n/arb/app_en.arb
@@ -988,8 +988,8 @@
   "cryptographicHolderBindingSubtitle": "Default : On\nDisable to accept Bearer credentials as tickets with low assurance.",
   "scopeParameters": "Scope Parameters",
   "scopeParametersSubtitle": "Default : Off\nEnable to force wallet to use scope instead of authorization_details.",
-  "preRegisteredWallet": "Pre registered wallet",
-  "preRegisteredWalletSubtitle": "Default : Off\nEnable to use client_id and client_secret values.",
+  "clientAuthentication": "Client Authentication",
+  "clientAuthenticationSubtitle": "Default : Public client\nEnable to use the Basic Authentication scheme.",
   "theServiceIsNotAvailable": "The service is not available",
   "issuerDID": "Issuer DID",
   "subjectDID": "Subject DID",
