+++ lib/l10n/untranslated.json
@@ -893,8 +893,8 @@
     "cryptographicHolderBindingSubtitle",
     "scopeParameters",
     "scopeParametersSubtitle",
-    "preRegisteredWallet",
-    "preRegisteredWalletSubtitle",
+    "clientAuthentication",
+    "clientAuthenticationSubtitle",
     "theServiceIsNotAvailable",
     "issuerDID",
     "subjectDID",
@@ -1806,8 +1806,8 @@
     "cryptographicHolderBindingSubtitle",
     "scopeParameters",
     "scopeParametersSubtitle",
-    "preRegisteredWallet",
-    "preRegisteredWalletSubtitle",
+    "clientAuthentication",
+    "clientAuthenticationSubtitle",
     "theServiceIsNotAvailable",
     "issuerDID",
     "subjectDID",
@@ -2022,8 +2022,8 @@
     "cryptographicHolderBindingSubtitle",
     "scopeParameters",
     "scopeParametersSubtitle",
-    "preRegisteredWallet",
-    "preRegisteredWalletSubtitle",
+    "clientAuthentication",
+    "clientAuthenticationSubtitle",
     "theServiceIsNotAvailable",
     "issuerDID",
     "subjectDID",
@@ -2935,8 +2935,8 @@
     "cryptographicHolderBindingSubtitle",
     "scopeParameters",
     "scopeParametersSubtitle",
-    "preRegisteredWallet",
-    "preRegisteredWalletSubtitle",
+    "clientAuthentication",
+    "clientAuthenticationSubtitle",
     "theServiceIsNotAvailable",
     "issuerDID",
     "subjectDID",
